pyproject.toml
[tool.poetry]
name = "audio-promt"
version = "0.1.0"
description = "App for audio promt analysis"
authors = ["Dmytro Andreichuk <dimaandr20@gmail.com>"]
packages = [{ include = "src" }] 
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
uvicorn = "^0.31.0"
pydantic = {extras = ["email"], version = "^2.9.2"}
python-dotenv = "^1.0.1"
alembic = "^1.13.3"
greenlet = "^3.1.1"
psycopg2 = "^2.9.9"
pydantic-settings = "^2.5.2"
aiofiles = "^24.1.0"
python-multipart = "^0.0.12"
psycopg2-binary = "^2.9.9"
speechrecognition = "^3.10.4"
python-jose = "^3.3.0"
passlib = {extras = ["argon2"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
